PrivremeniCommit
Ovo je privremeni commit kako bi se spremile promjene na github
