+++ FOIKnjiznica/FOIKnjiznica.sln
@@ -9,6 +9,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FOIKnjiznica", "FOIKnjiznic
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FOIKnjiznicaWebServis", "FOIKnjiznicaWebServis\FOIKnjiznicaWebServis.csproj", "{D704807F-4844-4DF1-B8E8-ECBBF8379E1C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PINModul", "PINModul\PINModul.csproj", "{1F8E9812-2E3C-45A1-8AD3-4DAA80BEB6DB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InterfaceModule", "InterfaceModule\InterfaceModule.csproj", "{98404996-8269-44CD-82EE-51FE5C83A78D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UzorakModul", "UzorakModul\UzorakModul.csproj", "{AF4E0BA0-3577-4088-B6CF-E5FBA64C4DAD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,6 +35,18 @@ Global
 		{D704807F-4844-4DF1-B8E8-ECBBF8379E1C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D704807F-4844-4DF1-B8E8-ECBBF8379E1C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D704807F-4844-4DF1-B8E8-ECBBF8379E1C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1F8E9812-2E3C-45A1-8AD3-4DAA80BEB6DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1F8E9812-2E3C-45A1-8AD3-4DAA80BEB6DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1F8E9812-2E3C-45A1-8AD3-4DAA80BEB6DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1F8E9812-2E3C-45A1-8AD3-4DAA80BEB6DB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{98404996-8269-44CD-82EE-51FE5C83A78D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{98404996-8269-44CD-82EE-51FE5C83A78D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{98404996-8269-44CD-82EE-51FE5C83A78D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{98404996-8269-44CD-82EE-51FE5C83A78D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AF4E0BA0-3577-4088-B6CF-E5FBA64C4DAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AF4E0BA0-3577-4088-B6CF-E5FBA64C4DAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AF4E0BA0-3577-4088-B6CF-E5FBA64C4DAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AF4E0BA0-3577-4088-B6CF-E5FBA64C4DAD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
