+++ FOIKnjiznica/FOIKnjiznica/FOIKnjiznica/FOIKnjiznica.csproj
@@ -12,12 +12,15 @@
     <PackageReference Include="Rg.Plugins.Popup" Version="1.2.0.223" />
     <PackageReference Include="Xam.Plugin.DeviceInfo" Version="4.1.1" />
     <PackageReference Include="Xamarin.FFImageLoading.Forms" Version="2.4.11.982" />
-    <PackageReference Include="Xamarin.Forms" Version="4.4.0.991477" />
+    <PackageReference Include="Xamarin.Forms" Version="4.4.0.991537" />
     <PackageReference Include="Xamarin.Essentials" Version="1.3.1" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\..\FOIKnjiznicaWebServis\FOIKnjiznicaWebServis.csproj" />
+    <ProjectReference Include="..\..\InterfaceModule\InterfaceModule.csproj" />
+    <ProjectReference Include="..\..\PINModul\PINModul.csproj" />
+    <ProjectReference Include="..\..\UzorakModul\UzorakModul.csproj" />
   </ItemGroup>
 
   <ItemGroup>
@@ -53,9 +56,6 @@
     <EmbeddedResource Update="MasterDetailPage1Master.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Update="PinPostavljanje.xaml">
-      <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
-    </EmbeddedResource>
     <EmbeddedResource Update="PopUpPages\PrekidRezervacijePopupPage.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
@@ -77,7 +77,7 @@
     <EmbeddedResource Update="Sadrzaj.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Update="UzorakPostavljanje.xaml">
+    <EmbeddedResource Update="StranicaPrijave.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
   </ItemGroup>
