+++ FOIKnjiznica/PINModul/PINUI.xaml
@@ -3,10 +3,8 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:ff="clr-namespace:FFImageLoading.Forms;assembly=FFImageLoading.Forms"
              mc:Ignorable="d"
-             Title="PinPostavljanje"
-             x:Class="FOIKnjiznica.PinPostavljanje">
+             x:Class="PINModul.PINUI">
     <ContentPage.Content>
 
         <StackLayout>
